Add script to automate Android emulator setup
Introduces a PowerShell script that automates the process of checking for prerequisites (Java, SDK tools), creating an Android Virtual Device (AVD) if it doesn't exist, and launching the emulator.

The script follows an "Android App Launch Protocol" and includes optimizations for emulator performance. It also provides a placeholder for launching the user's application once the emulator is ready.

Signed-off-by: Banee Ishaque K <Baneeishaque@gmail.com>